Ensure GPU num is an int even when running on CPU
Until now, when running on a CPU, start_heur_fn_runners set the list
of GPU numbers, which should be List[int], to [None].

This led to an error in load_heuristic_fn, which expects gpu_num to
be an int.

Changing [None] to [-1] covers this case.

Simplifying and reusing the function get_available_gpu_nums further
reduces some duplication.